changed a couple of listeners

+++ app/src/main/java/com/example/recipe_app/SignupActivity.java
@@ -11,15 +11,18 @@ import android.view.View;
 import android.widget.Button;
 
 
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 
+import com.google.android.gms.tasks.Task;
 import com.google.android.material.textfield.TextInputEditText;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -70,45 +73,48 @@ public class SignupActivity extends AppCompatActivity {
                                 @Override
                                 public void onSuccess(AuthResult authResult) {
                                     String uuid = authResult.getUser().getUid();
+                                    String email = authResult.getUser().getEmail();
                                     Map<String,Object> mapping = new HashMap<String,Object>();
-                                    mapping.put("uuid", uuid);
+                                    mapping.put("email", email);
 
                                     //TODO: Check later if null value in db causes problems
-                                    mapping.put("uploaded_recipes", null);
-
-                                    db.collection("user")
-                                            .add(mapping)
-                                            .addOnFailureListener(new OnFailureListener() {
+                                    mapping.put("uploaded_recipes", Arrays.asList());
+                                    //db.collection("users").document(uuid).set(mapping)
+                                    db.collection("users")
+                                            .document(uuid)
+                                            .set(mapping)
+                                            .addOnCompleteListener(new OnCompleteListener<Void>() {
                                                 @Override
-                                                public void onFailure(@NonNull Exception e) {
-                                                    mAuth.getCurrentUser().delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                        @Override
-                                                        public void onSuccess(Void unused) {
-                                                            new AlertDialog.Builder(SignupActivity.this)
-                                                                    .setTitle("Error")
-                                                                    .setMessage(e.getMessage())
-                                                                    .setPositiveButton("Okay", new DialogInterface.OnClickListener() {
-                                                                        @Override
-                                                                        public void onClick(DialogInterface dialog, int which) {
-                                                                            emailInput.setText("");
-                                                                            passwordInput.setText("");
-                                                                        }
-                                                                    })
-                                                                    .create()
-                                                                    .show();
-                                                        }
-                                                    });
-
-
-                                                }
-                                            })
-                                            .addOnSuccessListener(new OnSuccessListener<DocumentReference>() {
-                                                @Override
-                                                public void onSuccess(DocumentReference documentReference) {
-                                                    //TODO: Write a wrapper class for the documents to pass it along the activities
-                                                    startActivity(new Intent(SignupActivity.this, HomeActivity.class));
+                                                public void onComplete(@NonNull Task<Void> task) {
+                                                    if(!task.isSuccessful())
+                                                    {
+                                                        mAuth.getCurrentUser().delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                            @Override
+                                                            public void onSuccess(Void unused) {
+                                                                new AlertDialog.Builder(SignupActivity.this)
+                                                                        .setTitle("Error")
+                                                                        .setMessage("Error occured with signing up")
+                                                                        .setPositiveButton("Okay", new DialogInterface.OnClickListener() {
+                                                                            @Override
+                                                                            public void onClick(DialogInterface dialog, int which) {
+                                                                                emailInput.setText("");
+                                                                                passwordInput.setText("");
+                                                                            }
+                                                                        })
+                                                                        .create()
+                                                                        .show();
+                                                            }
+                                                        });
+                                                    }
+                                                    else
+                                                    {
+                                                        //TODO: Write a wrapper class for the documents to pass it along the activities
+                                                        startActivity(new Intent(SignupActivity.this, HomeActivity.class));
+                                                    }
                                                 }
                                             });
+
+
 
 
                                 }
